+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to `maduser/argon` will be documented in this file.
 
-## [Unreleased]
+## [1.1.2] - 2026-05-21
 
 ### Fixed